# api/openapi.yaml
openapi: 3.0.4
info:
  title: Participium - OpenAPI 1.0
  description: Participium Server based.
  termsOfService: https://swagger.io/terms/
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
  version: 1.0.12
...
paths:
...
  /admin/maintainers:
      post:
        tags:
          - Admin
        summary: Admin creates a maintainer
        requestBody:
          required: true
          content:
            application/json:
              schema:
                type: object
                properties:
                  name:
                    type: string
                  email:
                    type: string
                  password:
                    type: string
                  categories:
                    type: array
                    items:
                      $ref: "#/components/schemas/OfficeType"
                  active:
                    type: boolean
                required:
                  - name
                  - email
                  - password
                  - categories
        responses:
          "200":
            description: successful operation
            content:
              application/json:
                schema:
                  $ref: "#/components/schemas/Maintainer"
          default:
            description: Unexpected error
            content:
              application/json:
                schema:
                  $ref: "#/components/schemas/Error"
...
  /admin/maintainers/{id}:
      patch:
        tags:
          - Admin
        summary: Admin updates a maintainer
        parameters:
          - in: path
            name: id
            required: true
            schema:
              type: integer
        requestBody:
          required: true
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Maintainer"
        responses:
          "200":
            description: successful operation
            content:
              application/json:
                schema:
                  $ref: "#/components/schemas/Maintainer"
          default:
            description: Unexpected error
            content:
              application/json:
                schema:
                  $ref: "#/components/schemas/Error"
...
  /admin:
    post:
      tags:
        - Admin
      summary: Admins create new Officer
      operationId: createOfficer
      requestBody:
        description: Created user object
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Officer"
          application/xml:
            schema:
              $ref: "#/components/schemas/Officer"
          application/x-www-form-urlencoded:
            schema:
              $ref: "#/components/schemas/Officer"
      responses:
        "200":
          description: successful operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Officer"
            application/xml:
              schema:
                $ref: "#/components/schemas/Officer"
        default:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
  /Reports:
    post:
      tags:
        - Reports
      summary: User upload a new Report
      operationId: uploadReport
      requestBody:
        description: Created user object
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Report"
          application/xml:
            schema:
              $ref: "#/components/schemas/Report"
          application/x-www-form-urlencoded:
            schema:
              $ref: "#/components/schemas/Report"
      responses:
        "200":
          description: successful operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Report"
            application/xml:
              schema:
                $ref: "#/components/schemas/Report"
        default:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    get:
      tags:
        - Reports
      summary: Retrieve all public Reports for map visualization
      description: Get all approved Reports to display on the public map
      operationId: getReports
      parameters: []
      responses:
        "200":
          description: successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Report"
        default:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
  /reports/{id}/message:
    post:
      tags:
        - Reports
      summary: Officer sends a message to report author
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                message:
                  type: string
                  example: "Aggiornamento sul tuo report"
              required:
                - message
      responses:
        "201":
          description: successful operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Notification"
        default:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /admin/role/add:
    patch:
      tags:
        - Admin
      summary: Admin updates officer roles
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                officerId:
                  type: integer
                roles:
                  type: array
                  items:
                    type: object
                    properties:
                      officerRole:
                        $ref: '#/components/schemas/OfficerRole'
                      officeType:
                        $ref: '#/components/schemas/OfficeType'
              required:
                - officerId
                - roles
      responses:
        "200":
          description: successful operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Officer"
        default:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /admin/role/remove:
    patch:
      tags:
        - Admin
      summary: Admin updates officer roles
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                officerId:
                  type: integer
                roles:
                  type: array
                  items:
                    type: object
                    properties:
                      officerRole:
                        $ref: '#/components/schemas/OfficerRole'
                      officeType:
                        $ref: '#/components/schemas/OfficeType'
              required:
                - officerId
                - roles
      responses:
        "200":
          description: successful operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Officer"
        default:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
components:
  schemas:
    OfficeType:
      type: string
      enum: [ "infrastructure", "environment","safety", "sanitation","transport","organization","other"]
    OfficerRole:
      type: string
      enum: ["municipal_public_relations_officer", "municipal_administrator","technical_office_staff"]
...
    User:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 2
        username:
          type: string
          example: username
        firstName:
          type: string
          example: Mario
        lastName:
          type: string
          example: Rossi
        password:
          type: string
          example: placeholder
        email:
          type: string
          example: mail@domain.com
        avatar:
          type: string
          example: /uploads/avatars/167123123-avatar.jpg
        telegramUsername:
          type: string
          example: my_telegram
        emailNotifications:
          type: boolean
          example: true
      xml:
        name: user
    Officer:
      type: object
      properties:
        id:
          type: integer
        username:
          type: string
        name:
          type: string
        surname:
          type: string
        email:
          type: string
        roles:  # Array invece di singolo
          type: array
          items:
            type: object
            properties:
              officerRole:
                $ref: '#/components/schemas/OfficerRole'
              officeType:
                $ref: '#/components/schemas/OfficeType'
    Location:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        Coordinates:
          type: object
          properties:
            longitude:
              type: number
            latitude:
              type: number
    Report:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        location:
          $ref: "#/components/schemas/Location"
        Author:
          $ref: "#/components/schemas/User"
        Anonymity:
          type: boolean
        Date:
          type: string
        Category:
          $ref: "#/components/schemas/OfficeType"
        Document:
          type: object
          properties:
            Description:
              type: string
              minLength: 150
              example: the water is clear
            Photos:
              type: array
              items:
                type: string
              minItems: 1
              maxItems: 3
        assignedMaintainerId:
          type: integer
          nullable: true
...
    Notification:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        reportId:
          type: integer
        type:
          type: string
          enum:
            - STATUS_CHANGE
            - OFFICER_MESSAGE
        message:
          type: string
        createdAt:
          type: string
          format: date-time
        read:
          type: boolean
    Maintainer:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        password:
          type: string
        categories:
          type: array
          items:
            $ref: "#/components/schemas/OfficeType"
        active:
          type: boolean
...
    Error:
      type: object
      properties:
        code:
          type: string
        message:
          type: string
      required:
        - code
        - message
...
  securitySchemes:
    petstore_auth:
      type: oauth2
      flows:
        implicit:
          authorizationUrl: https://petstore3.swagger.io/oauth/authorize
          scopes:
            write:pets: modify pets in your account
            read:pets: read your pets
    api_key:
      type: apiKey
      name: api_key
      in: header